+++ mmdet/core/bbox/coder/tblr_bbox_coder.py
@@ -184,7 +184,7 @@ def tblr2bboxes(priors,
     if clip_border and max_shape is not None:
         # clip bboxes with dynamic `min` and `max` for onnx
         if torch.onnx.is_in_onnx_export():
-            from mmdet.core.export.onnx_helper import dynamic_clip_for_onnx
+            from mmdet.core.export import dynamic_clip_for_onnx
             xmin, ymin, xmax, ymax = dynamic_clip_for_onnx(
                 xmin, ymin, xmax, ymax, max_shape)
             bboxes = torch.cat([xmin, ymin, xmax, ymax], dim=-1)
