+++ mmdet/core/export/test.py
@@ -1,0 +1,106 @@
+import os.path as osp
+import warnings
+
+import numpy as np
+import onnxruntime as ort
+import torch
+
+from mmdet.core import bbox2result
+from mmdet.models import BaseDetector
+
+
+class ONNXRuntimeDetector(BaseDetector):
+    """Wrapper for detector's inference with ONNXRuntime."""
+
+    def __init__(self, onnx_file, class_names, device_id):
+        super(ONNXRuntimeDetector, self).__init__()
+        # get the custom op path
+        ort_custom_op_path = ''
+        try:
+            from mmcv.ops import get_onnxruntime_op_path
+            ort_custom_op_path = get_onnxruntime_op_path()
+        except (ImportError, ModuleNotFoundError):
+            warnings.warn('If input model has custom op from mmcv, \
+                you may have to build mmcv with ONNXRuntime from source.')
+        session_options = ort.SessionOptions()
+        # register custom op for onnxruntime
+        if osp.exists(ort_custom_op_path):
+            session_options.register_custom_ops_library(ort_custom_op_path)
+        sess = ort.InferenceSession(onnx_file, session_options)
+        providers = ['CPUExecutionProvider']
+        options = [{}]
+        is_cuda_available = ort.get_device() == 'GPU'
+        if is_cuda_available:
+            providers.append('CUDAExecutionProvider')
+            options.append({'device_id': device_id})
+
+        sess.set_providers(providers, options)
+
+        self.sess = sess
+        self.CLASSES = class_names
+        self.device_id = device_id
+        self.io_binding = sess.io_binding()
+        self.output_names = [_.name for _ in sess.get_outputs()]
+        self.is_cuda_available = is_cuda_available
+
+    def simple_test(self, img, img_metas, **kwargs):
+        raise NotImplementedError('This method is not implemented.')
+
+    def aug_test(self, imgs, img_metas, **kwargs):
+        raise NotImplementedError('This method is not implemented.')
+
+    def extract_feat(self, imgs):
+        raise NotImplementedError('This method is not implemented.')
+
+    def forward_test(self, imgs, img_metas, **kwargs):
+        input_data = imgs[0]
+        img_metas = img_metas[0]
+        batch_size = input_data.shape[0]
+        # set io binding for inputs/outputs
+        if self.is_cuda_available:
+            self.io_binding.bind_input(
+                name='input',
+                device_type='cuda',
+                device_id=self.device_id,
+                element_type=np.float32,
+                shape=input_data.shape,
+                buffer_ptr=input_data.data_ptr())
+        else:
+            self.io_binding.bind_cpu_input('input', input_data.cpu().numpy())
+
+        for name in self.output_names:
+            self.io_binding.bind_output(name)
+        # run session to get outputs
+        self.sess.run_with_iobinding(self.io_binding)
+        ort_outputs = self.io_binding.copy_outputs_to_cpu()
+        batch_dets, batch_labels = ort_outputs[:2]
+        batch_masks = ort_outputs[2] if len(ort_outputs) == 3 else None
+
+        results = []
+        for i in range(batch_size):
+            scale_factor = img_metas[i]['scale_factor']
+            dets, labels = batch_dets[i], batch_labels[i]
+            dets[:, :4] /= scale_factor
+            dets_results = bbox2result(dets, labels, len(self.CLASSES))
+            if batch_masks is not None:
+                masks = batch_masks[i]
+                img_h, img_w = img_metas[i]['img_shape'][:2]
+                ori_h, ori_w = img_metas[i]['ori_shape'][:2]
+                masks = masks[:, :img_h, :img_w]
+                mask_dtype = masks.dtype
+                masks = masks.astype(np.float32)
+                masks = torch.from_numpy(masks)
+                masks = torch.nn.functional.interpolate(
+                    masks.unsqueeze(0), size=(ori_h, ori_w))
+                masks = masks.squeeze(0).detach().numpy()
+                # convert mask to range(0,1)
+                if mask_dtype != np.bool:
+                    masks /= 255
+                masks = masks >= 0.5
+                segms_results = [[] for _ in range(len(self.CLASSES))]
+                for j in range(len(dets)):
+                    segms_results[labels[j]].append(masks[j])
+                results.append((dets_results, segms_results))
+            else:
+                results.append(dets_results)
+        return results
